pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pharmacast-ai"
version = "0.1.0"
description = "Shared Python services for PharmaForecast forecasting and LLM workflows."
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.110,<1.0",
  "numpy>=1.26,<3.0",
  "pandas>=2.0,<3.0",
  "prophet>=1.1,<2.0",
  "pydantic-settings>=2.0,<3.0",
  "python-dateutil>=2.8,<3.0",
  "supabase>=2.0,<3.0",
  "uvicorn[standard]>=0.27,<1.0",
]

[project.optional-dependencies]
test = [
  "httpx>=0.27,<1.0",
  "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "shared*", "ai*", "forecasting*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
